Adjusts server_url for fetching charts
Currently pulling a tgz in from a normal chartmuseum produces an empty
tarball.

This is a terrible "fix", but it seems that server_url cannot be a base
url because of the differences in path between Harbor and normal
Chartmuseum, we need the user to tell us about the full path to the api.

Harbor apparently decided that to change functionality between metadata
and modification/fetching we need to modify this url not in some sane
way, like entirely on the right side of the path after the endpoint, but
on two different paths, ie:
/api/<somearbitrarynoise>/charts and /<somearbitrarynoise>/charts.

https://github.com/helm/chartmuseum/blob/b58d67830e793bac2a345aade98c36a7825104fa/README.md#api